fux: refactor
