+++ lib/main.dart
@@ -1,33 +1,6 @@
 import 'package:flutter/material.dart';
-
-import 'package:ninety9names/pages/home/home.dart';
-import 'package:ninety9names/pages/names/details.dart';
-
+import 'package:ninety9names/pages/root/ninety9names.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(Ninety9Names());
 }
-
-class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-        title: 'ninety9names',
-        theme: ThemeData(
-          primarySwatch: Colors.green,
-          visualDensity: VisualDensity.adaptivePlatformDensity,
-        ),
-        initialRoute: '/',
-        routes: {
-          '/': (context) => MyHomePage(title: 'ninety9names'),
-          '/details': (context) => Details(),
-        });
-  }
-}
-
-
-
-
-
-
